Team- and thread-level sort, sort_by_key
Add sort functions that can be called from device,
and exploit team and thread level parallelism. The new functions
use bitonic sort, which is good for this because it's in-place but
highly parallel (when sorting N items, N/2 pairs are compared at once).
It's also comparison-based, so there are versions that can accept an
arbitrary comparison functor (operator()(a, b) returns true if key a
goes before key b).

sort_by_key is the same, but in addition to the keys it takes a values
view of the same length. The pairs keys(i) and values(i) are all sorted
according to the key. This is useful for sorting CRS matrices for
example.